Update timer interface

+++ src/framework.c
@@ -97,7 +97,7 @@ void framework_mousereleased(int x, int y, const char *button, int istouch) {
 void framework_wheelmoved(int x, int y) {
 }
 
-void framework_update(double dt) {
+void framework_update(uint64_t dt) {
 }
 
 void framework_draw() {
